Import Cobalt 6.16000 -> 9.28152

+++ src/cobalt/cssom/shadow_value.h
@@ -1,18 +1,16 @@
-/*
- * Copyright 2015 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2015 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #ifndef COBALT_CSSOM_SHADOW_VALUE_H_
 #define COBALT_CSSOM_SHADOW_VALUE_H_
@@ -45,7 +43,7 @@ class ShadowValue : public PropertyValue {
   ShadowValue(const std::vector<scoped_refptr<LengthValue> >& lengths,
               const scoped_refptr<RGBAColorValue>& color, bool has_inset)
       : color_(color), has_inset_(has_inset) {
-    DCHECK_LE(lengths.size(), kMaxLengths);
+    DCHECK_LE(lengths.size(), static_cast<size_t>(kMaxLengths));
     for (size_t i = 0; i < lengths.size(); ++i) {
       lengths_[i] = lengths[i];
     }
